Proyecto Finalizado, implementacion de auth

+++ src/routes/links.js
@@ -1,18 +1,20 @@
 const express = require('express');
 const router = express.Router();
 const pool = require('../database');
+const { isLoggedIn } = require('../lib/auth');
 
-router.get('/add', (req, res) => {
+router.get('/add', isLoggedIn, (req, res) => {
 	res.render('links/add');
 });
 
 //Agregar un link
-router.post('/add', async (req, res) => {
+router.post('/add', isLoggedIn, async (req, res) => {
 	const { title, url, description } = req.body;
 	const newLink = {
 		title,
 		url,
-		description
+		description,
+		user_id     : req.user.id
 	};
 	await pool.query('insert into links set ?', [ newLink ]);
 	req.flash('success', 'Link agregado con exito!');
@@ -20,21 +22,21 @@ router.post('/add', async (req, res) => {
 });
 
 //Listar todos los links
-router.get('/', async (req, res) => {
-	const links = await pool.query('select * from links');
+router.get('/', isLoggedIn, async (req, res) => {
+	const links = await pool.query('select * from links where user_id = ?', [ req.user.id ]);
 	res.render('links/links', { links });
 });
 
 //Eliminar un Link
-router.get('/delete/:id', async (req, res) => {
+router.get('/delete/:id', isLoggedIn, async (req, res) => {
 	const { id } = req.params;
-	await pool.query('delete from links where id = ?', [ id ]);
+	await pool.query('delete from links where id = ? and user_id = ?', [ id, req.user.id ]);
 	req.flash('success', 'Link eliminado con exito!');
 	res.redirect('/links');
 });
 
 //Editar Link
-router.post('/edit/:id', async (req, res) => {
+router.post('/edit/:id', isLoggedIn, async (req, res) => {
 	const { id } = req.params;
 	const { title, url, description } = req.body;
 	updatedLink = {
@@ -49,7 +51,7 @@ router.post('/edit/:id', async (req, res) => {
 });
 
 //Editar link
-router.get('/edit/:id', async (req, res) => {
+router.get('/edit/:id', isLoggedIn, async (req, res) => {
 	const { id } = req.params;
 	const link = await pool.query('select * from links where id = ?', [ id ]);
 	res.render('links/edit', { link: link[0] });
